+++ UiBundle/src/Form/Block/CollectionType.php
@@ -10,6 +10,7 @@
 
 namespace c975L\UiBundle\Form\Block;
 
+use c975L\UiBundle\Form\LinkTargetType;
 use c975L\UiBundle\Registry\CollectionSourceRegistry;
 use c975L\UiBundle\Service\BlockAnchorSlugger;
 use Symfony\Component\Form\AbstractType;
@@ -85,7 +86,7 @@ class CollectionType extends AbstractType
                 'label' => 'label.link_label',
                 'required' => false,
             ])
-            ->add('linkUrl', TextType::class, [
+            ->add('linkUrl', LinkTargetType::class, [
                 'label' => 'label.url',
                 'required' => false,
             ])