+++ UiBundle/src/c975LUiBundle.php
@@ -32,6 +32,7 @@ use c975L\UiBundle\DependencyInjection\Compiler\FormPageUrlProviderPass;
 use c975L\UiBundle\DependencyInjection\Compiler\FormThemeRegistryPass;
 use c975L\UiBundle\DependencyInjection\Compiler\GalleryShowcaseProviderPass;
 use c975L\UiBundle\DependencyInjection\Compiler\InternalLinkLocalizerPass;
+use c975L\UiBundle\DependencyInjection\Compiler\LinkTargetProviderPass;
 use c975L\UiBundle\DependencyInjection\Compiler\MediaUsageProviderPass;
 use c975L\UiBundle\DependencyInjection\Compiler\PdfDocumentSourcePass;
 use c975L\UiBundle\DependencyInjection\Compiler\PlaceholderMediaProviderPass;
@@ -85,6 +86,7 @@ class c975LUiBundle extends AbstractBundle
         $container->addCompilerPass(new EmailLayoutProviderPass());
         $container->addCompilerPass(new EmailAttachmentProviderPass());
         $container->addCompilerPass(new InternalLinkLocalizerPass());
+        $container->addCompilerPass(new LinkTargetProviderPass());
         $container->addCompilerPass(new FontProviderPass());
         $container->addCompilerPass(new FormThemeRegistryPass());
         $container->addCompilerPass(new FormActionProviderPass());