+++ UiBundle/tests/Templates/CollectionItemPortraitVariantTest.php
@@ -11,6 +11,8 @@
 namespace c975L\UiBundle\Tests\Templates;
 
 use PHPUnit\Framework\TestCase;
+use Symfony\Bridge\Twig\Extension\TranslationExtension;
+use Symfony\Component\Translation\IdentityTranslator;
 use Twig\Environment;
 use Twig\Loader\FilesystemLoader;
 
@@ -96,6 +98,8 @@ class CollectionItemPortraitVariantTest extends TestCase
     private function render(array $context): string
     {
         $twig = new Environment(new FilesystemLoader(\dirname(__DIR__, 2) . '/templates'));
+        // Untranslated keys come back as-is, which is enough for the button's generic label
+        $twig->addExtension(new TranslationExtension(new IdentityTranslator()));
 
         return $twig->render('blocks/CollectionItem.html.twig', $context);
     }